Changed gen_mksurfdata CMakeLists.txt to use libpio .a files instead of .so files.
The ParallelIO modules on betzy don't include shared library files
libpiof.so libpioc.so, and instead only have statically linked
libraries libpiof.a and libpioc.a. This commit simply changes the
filenames to libpiof.a and libpioc.a in the text file used to generate
the makefiles to build mksurfdata_esmf. After this change, the
mksurfdata_esmf executable compiles successfully, but (**NB!**) it's not
clear yet whether it will run, or produce correct results.