Revert "Added the ctsm python package to the pixi env"
This reverts commit 7e91d4554c08a923e99a0235665ecf2385d4aa3c.

It shouldn't really be necessary to install the ctsm Python package,
since the scripts that use it add the required directory to sys.path
before importing. But the reverted commit might still be useful in a
situation where we need to use the ctsm package as part of other Python
code, and don't want to manually add to sys.path.