Preliminary instructions for preparing mksurfdata run.
Added preliminary instructions for building the mksurfdata executable
and for creating the mksurfdata namelist file.

Note that some postprocessing will be needed on the namelist file, and
the README will need to be updated to include instructions for that.
Just running gen_mksurfdata_namelist gives a namlist where 4 of the
input data file paths point to `/glade/campaign` (presumably a
hardcoding bug in the gen_mksurfdata_namelist script), and the specified
machine name needs to be changed from the name of the login node to the
cluster (e.g., from `login-1` to `betzy`).