Renamed mksurfdata jobscript to make clear that it's specific.
The jobscript is specific to betzy, and uses hardcoded paths for the
user janko. Must be adapted for other users and/or machines and repo
locations. `gen_mksurfdata_jobscript_single` doesn't work for this
(except as a starting point that must be heavily edited), since it only
supports NCAR machines.