Generated new namelists and successfully downloaded input files for mksurfdata.
This had to be done using a temporary hack to get around permission
errors in the /cluster/shared/noresm/inputdata/lnd/clm2/rawdata
directory. The directory was moved out of the way, and a new empty directory
was temporarily used and then renamed. Will need to change the
mksurfdata namelist to use the renamed name for the rawdata directory.