Removed upper-version pins and rebuilt pixi Python environment.
Some problem was causing builds and even ipython startups to crash.
Seemed related to the pypi dependencies (Sphinx?) in the pixi
environment. Was solved by removing all existing environments, cleaning
the cache completely, removing the pixi.lock file, and reinstalling
environments through `pixi install`. Not clear which ones of these were
necessary, but several of them did not work isolation. Also not clear
whether the problem was that something broke in the environment or pixi
cache itself, or whether there was a change on betzy.