Fixed bugs in how NetCDF mesh files are written.
This affects both the temporary file that is written when creating an esmpy.Mesh
object, and when writing the final mesh file with land mask and element areas
added. Xarray does not currently (version 2025.12.1) support writing Datasets
with MultiIndex indexes to NetCDF files, so these need to be reset before
writing both for the temporary file and for the final result (the latter should
be done anyway, to conform to the standard ESMF mesh file format). Also, the
temporary file used for the area calculation needs to be clobbered, since it is
created as an empty file when obtaining a temporary file name from the tempfile
module.